fix(macos): move install-location state into its own leaf module
`sentry.ts` reading the outcome from `move-to-applications.ts` pulled
`BrowserWindow` into the Sentry module's import graph, which broke
`sentry.test.ts` (its electron mock supplies only `app`). The layering was
wrong either way: a telemetry module should not import the installer.

Hold the value in `install-location.ts`, which touches only `app`. The
installer records into it and Sentry reads from it, so neither depends on
the other.
