+++ clients/macos/src/main/move-to-applications.ts
@@ -1,5 +1,6 @@
 import { app, BrowserWindow } from "electron";
 
+import { recordInstallLocation } from "./install-location";
 import log from "./logger";
 
 // Build-time define (see electron.vite.config.ts) — the same VELLUM_ENVIRONMENT
@@ -104,46 +105,13 @@ function createInstallSplash(): BrowserWindow {
  *
  * @see https://www.electronjs.org/docs/latest/api/app#appmovetoapplicationsfolderoptions-macos
  */
-/**
- * Where the running app ended up, and for the negative cases which branch put
- * it there. A packaged app outside /Applications keeps running from a
- * read-only location (a mounted DMG, a quarantined ~/Downloads copy), where
- * `electron-updater` refuses every update with "Cannot update while running on
- * a read-only volume". The branches are indistinguishable from the error alone,
- * so `sentry.ts` reports this as the `install_location` tag.
- */
-export type InstallLocation =
-  | "unpackaged"
-  | "applications"
-  | "relocating"
-  | "skipped-pending-open"
-  | "conflict-exists-and-running"
-  | "declined"
-  | "failed";
-
-let installLocation: InstallLocation = "unpackaged";
-
-export const getInstallLocation = (): InstallLocation => installLocation;
-
-/**
- * Record that a launch carrying a `.vellum` file or a deep link bypassed the
- * relocation, so a packaged app left outside /Applications is attributable to
- * that skip rather than to a failed move.
- */
-export const markRelocationSkipped = (): void => {
-  if (!app.isPackaged) return;
-  installLocation = app.isInApplicationsFolder()
-    ? "applications"
-    : "skipped-pending-open";
-};
-
 export async function relocateToApplicationsFolder(): Promise<boolean> {
   if (!app.isPackaged) {
-    installLocation = "unpackaged";
+    recordInstallLocation("unpackaged");
     return false;
   }
   if (app.isInApplicationsFolder()) {
-    installLocation = "applications";
+    recordInstallLocation("applications");
     return false;
   }
 
@@ -194,17 +162,19 @@ export async function relocateToApplicationsFolder(): Promise<boolean> {
     });
     // A `false` return is the user's own doing (the conflict handler above, or
     // a cancelled authorization prompt); anything else throws.
-    installLocation = moved
-      ? "relocating"
-      : conflictedWithRunningCopy
-        ? "conflict-exists-and-running"
-        : "declined";
+    recordInstallLocation(
+      moved
+        ? "relocating"
+        : conflictedWithRunningCopy
+          ? "conflict-exists-and-running"
+          : "declined",
+    );
     // On success the process is already quitting/relaunching, so the splash
     // is torn down with it; only close it if we're staying put.
     if (!moved) closeSplash();
     return moved;
   } catch (err) {
-    installLocation = "failed";
+    recordInstallLocation("failed");
     log.error("[move-to-applications] moveToApplicationsFolder failed:", err);
     closeSplash();
     return false;
@@ -214,5 +184,5 @@ export async function relocateToApplicationsFolder(): Promise<boolean> {
 // Test seam, exported only for unit-test setup. Production code never
 // resets the recorded location.
 export const __resetForTesting = (): void => {
-  installLocation = "unpackaged";
+  recordInstallLocation("unpackaged");
 };
